html/semantics/scripting-1/the-script-element/json-module/json-module-service-worker-test.https.tentative.html

<!doctype html>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
    promise_test(async (test) => {
        const reg = await navigator.serviceWorker.register('./serviceworker.js', { type: 'module' })
        test.add_cleanup(() => reg.unregister());
        assert_not_equals(reg.installing, undefined);
    }, "Javascript importing JSON Module should load within the context of a service worker");

    promise_test(async (test) => {
        const reg = await navigator.serviceWorker.register('./module.json', { type: 'module' })
        test.add_cleanup(() => reg.unregister());
        assert_not_equals(reg.installing, undefined);
    }, "JSON Modules should load within the context of a service worker");

    promise_test(async (test) => {
        const reg = await navigator.serviceWorker.register('./serviceworker-dynamic-import.js', { type: 'module' })
        test.add_cleanup(() => reg.unregister());
       reg.onmessage = (e) => {
            assert_not_equals(reg.installing, undefined);
            assert_equals(e.data, "LOADED")
        };
    }, "JSON Module dynamic import should load within the context of a service worker");
</script>
